cleanup: remove mixed modules warning
