+++ clients/trieve-shopify-extension/app/dehydrate.ts
@@ -1,0 +1,19 @@
+import { useMatches } from "@remix-run/react";
+import merge from "deepmerge";
+import { DehydratedState } from "@tanstack/react-query";
+
+export const useDehydratedState = (): DehydratedState => {
+  const matches = useMatches();
+
+  const dehydratedState = matches
+    // @ts-ignore
+    .map((match) => match.data?.dehydratedState)
+    .filter(Boolean);
+
+  return dehydratedState.length
+    ? dehydratedState.reduce(
+        (accumulator, currentValue) => merge(accumulator, currentValue),
+        {},
+      )
+    : undefined;
+};
