+++ clients/trieve-shopify-extension/app/serverQuery.ts
@@ -1,9 +1,7 @@
-import merge from "deepmerge";
 import { json, LoaderFunctionArgs } from "@remix-run/node";
-import { dehydrate, DehydratedState, QueryClient } from "@tanstack/react-query";
+import { dehydrate, QueryClient } from "@tanstack/react-query";
 import { TrieveSDK } from "trieve-ts-sdk";
 import { sdkFromKey, validateTrieveAuth } from "./auth";
-import { useMatches } from "@remix-run/react";
 
 interface TrieveServerLoaderCtx {
   queryClient: QueryClient;
@@ -22,20 +20,4 @@ export const createServerLoader = (
       dehydratedState: dehydrate(queryClient),
     });
   };
-};
-
-export const useDehydratedState = (): DehydratedState => {
-  const matches = useMatches();
-
-  const dehydratedState = matches
-    // @ts-ignore
-    .map((match) => match.data?.dehydratedState)
-    .filter(Boolean);
-
-  return dehydratedState.length
-    ? dehydratedState.reduce(
-        (accumulator, currentValue) => merge(accumulator, currentValue),
-        {},
-      )
-    : undefined;
 };
